feat: implement STRUCT type support for PyAthena SQLAlchemy dialect
This commit adds comprehensive STRUCT type support to PyAthena, addressing GitHub issue #454.

Changes include:
- Add struct converter function in converter.py for JSON-to-dict conversion
- Implement AthenaStruct type class with field definitions and SQLAlchemy integration
- Add struct compilation support in type compiler with visit_struct/visit_STRUCT methods
- Refactor code organization by moving compiler classes to compiler.py
- Move identifier preparer classes to preparer.py for better separation of concerns
- Add comprehensive test coverage for all struct functionality
- Update ischema_names mapping to recognize struct/row types

Benefits:
- Enables querying and manipulating STRUCT/ROW data types in Athena
- Provides type-safe field access and validation
- Maintains full backward compatibility with existing code
- Follows PyAthena's architectural patterns and DB API 2.0 compliance
