+++ pyathena/sqlalchemy/types.py
@@ -2,7 +2,7 @@
 from __future__ import annotations
 
 from datetime import date, datetime
-from typing import TYPE_CHECKING, Any, Optional, Union
+from typing import TYPE_CHECKING, Any, Dict, Optional, Tuple, Union
 
 from sqlalchemy.sql import sqltypes
 from sqlalchemy.sql.type_api import TypeEngine
@@ -46,3 +46,34 @@ class Tinyint(sqltypes.Integer):
 
 class TINYINT(Tinyint):
     __visit_name__ = "TINYINT"
+
+
+class AthenaStruct(TypeEngine[Dict[str, Any]]):
+    __visit_name__ = "struct"
+
+    def __init__(self, *fields: Union[str, Tuple[str, Any]]) -> None:
+        self.fields: Dict[str, TypeEngine[Any]] = {}
+
+        for field in fields:
+            if isinstance(field, str):
+                self.fields[field] = sqltypes.String()
+            elif isinstance(field, tuple) and len(field) == 2:
+                field_name, field_type = field
+                if isinstance(field_type, TypeEngine):
+                    self.fields[field_name] = field_type
+                else:
+                    # Assume it's a SQLAlchemy type class and instantiate it
+                    self.fields[field_name] = field_type()
+            else:
+                raise ValueError(f"Invalid field specification: {field}")
+
+    def __getitem__(self, key: str) -> TypeEngine[Any]:
+        return self.fields[key]
+
+    @property
+    def python_type(self) -> type:
+        return dict
+
+
+class STRUCT(AthenaStruct):
+    __visit_name__ = "STRUCT"
